Changed file structure and completed home page
Added a separate folder "homepage" in components that has all the components that exist in the homepage.

Changed all pressables to touchable opacities.

Made the daily goals scrollable.
Added the icons for the rest of the daily goals and updated the tasks text.

Separated the Content for you into a separate files.

Separated the Suggested tools into a separate file.

Completed the Suggested  tools

Added the Content for you images and text.